organize graph-view related parts into a src subdirectory
