+++ src/graph-view/start.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="./graph/graph"></require>
+  <require from="../graph/graph"></require>
   <require from="source-code"></require>
 
   <!-- Search -->
